refactor(card-browser): move 'toggle*' to Fragment
* toggleMark
* toggleSuspendCards
* toggleBury
