+++ AnkiDroid/src/main/java/com/ichi2/anki/CardBrowser.kt
@@ -74,6 +74,9 @@ import com.ichi2.anki.browser.SharedPreferencesLastDeckIdRepository
 import com.ichi2.anki.browser.registerFindReplaceHandler
 import com.ichi2.anki.browser.showChangeDeckDialog
 import com.ichi2.anki.browser.toCardBrowserLaunchOptions
+import com.ichi2.anki.browser.toggleBury
+import com.ichi2.anki.browser.toggleMark
+import com.ichi2.anki.browser.toggleSuspendCards
 import com.ichi2.anki.common.annotations.NeedsTest
 import com.ichi2.anki.common.utils.annotation.KotlinCleanup
 import com.ichi2.anki.dialogs.BrowserOptionsDialog
@@ -844,16 +847,6 @@ open class CardBrowser :
             }
         updateFlagForSelectedRows(flag)
     }
-
-    /** All the notes of the selected cards will be marked
-     * If one or more card is unmarked, all will be marked,
-     * otherwise, they will be unmarked  */
-    @NeedsTest("Test that the mark get toggled as expected for a list of selected cards")
-    @VisibleForTesting
-    fun toggleMark() =
-        launchCatchingTask {
-            withProgress { viewModel.toggleMark() }
-        }
 
     /** Opens the note editor for a card.
      * We use the Card ID to specify the preview target  */
@@ -1723,21 +1716,6 @@ open class CardBrowser :
     ) {
         updateList()
     }
-
-    private fun toggleSuspendCards() = launchCatchingTask { withProgress { viewModel.toggleSuspendCards().join() } }
-
-    /** @see CardBrowserViewModel.toggleBury */
-    private fun toggleBury() =
-        launchCatchingTask {
-            val result = withProgress { viewModel.toggleBury() } ?: return@launchCatchingTask
-            // show a snackbar as there's currently no colored background for buried cards
-            val message =
-                when (result.wasBuried) {
-                    true -> TR.studyingCardsBuried(result.count)
-                    false -> resources.getQuantityString(R.plurals.unbury_cards_feedback, result.count, result.count)
-                }
-            showUndoSnackbar(message)
-        }
 
     fun showUndoSnackbar(message: CharSequence) {
         showSnackbar(message) {
